+++ RevitAddin.DA.Tester/Revit/App.cs
@@ -1,6 +1,5 @@
 ﻿using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.DB;
-using DesignAutomationFramework;
 using RevitAddin.DA.Tester.Services;
 using System;
 
@@ -8,6 +7,7 @@ namespace RevitAddin.DA.Tester.Revit
 {
     public class App : IExternalDBApplication
     {
+        IDisposable designAutomation;
         public ExternalDBApplicationResult OnStartup(ControlledApplication application)
         {
             Console.WriteLine("----------------------------------------");
@@ -18,22 +18,14 @@ namespace RevitAddin.DA.Tester.Revit
             Console.WriteLine($"AddInName: \t{application.ActiveAddInId?.GetAddInName()}");
             Console.WriteLine("----------------------------------------");
 
-            DesignAutomationBridge.DesignAutomationReadyEvent += DesignAutomationBridge_DesignAutomationReadyEvent;
+            designAutomation = new DesignAutomationLoadVersion(typeof(DesignAutomationController));
             return ExternalDBApplicationResult.Succeeded;
         }
 
         public ExternalDBApplicationResult OnShutdown(ControlledApplication application)
         {
-            DesignAutomationBridge.DesignAutomationReadyEvent -= DesignAutomationBridge_DesignAutomationReadyEvent;
+            designAutomation?.Dispose();
             return ExternalDBApplicationResult.Succeeded;
-        }
-
-        private void DesignAutomationBridge_DesignAutomationReadyEvent(object sender, DesignAutomationReadyEventArgs e)
-        {
-            DesignAutomationBridge.DesignAutomationReadyEvent -= DesignAutomationBridge_DesignAutomationReadyEvent;
-
-            var data = e.DesignAutomationData;
-            e.Succeeded = DesignAutomationController.Execute(data.RevitApp, data.FilePath, data.RevitDoc);
         }
     }
 }