+++ RevitAddin.DA.Tester/Services/DesignAutomationController.cs
@@ -1,6 +1,7 @@
 ﻿using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.DB;
 using RevitAddin.DA.Tester.Models;
+using RevitAddin.DA.Tester.Revit;
 using System;
 using System.Linq;
 using System.Reflection;
@@ -8,9 +9,9 @@ using System.Runtime.Versioning;
 
 namespace RevitAddin.DA.Tester.Services
 {
-    public class DesignAutomationController
+    public class DesignAutomationController : IDesignAutomation
     {
-        public static bool Execute(Application application, string filePath = null, Document document = null)
+        public bool Execute(Application application, string filePath = null, Document document = null)
         {
             var inputModel = new InputModel().Load();